refactor: move mock functions to class recipe controller
